+++ includes/modules/payment/easycredit.php
@@ -38,7 +38,7 @@
   
       if ($this->enabled === true) {
         if ((int) MODULE_PAYMENT_EASYCREDIT_ORDER_STATUS_ID > 0) {
-          $this->order_status = MODULE_PAYMENT_EASYCREDIT_ORDER_STATUS_ID;
+          $this->order_status = (int) MODULE_PAYMENT_EASYCREDIT_ORDER_STATUS_ID;
         }
         if ((int) MODULE_PAYMENT_EASYCREDIT_ORDER_STATUS_SUCCESS_ID > 0) {
           $this->order_status_success = MODULE_PAYMENT_EASYCREDIT_ORDER_STATUS_SUCCESS_ID;
